requirements.txt
openai-agents>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncio
pydantic>=2.0.0 